requirements.txt
numpy==1.23.1
Pillow==9.2.0
qrcode==7.3.1
streamlit==1.12.2
